requirements.txt
accelerate==0.29.3
aiofiles==23.1.0
anyio==4.3.0
attrs==22.2.0
beautifulsoup4==4.12.3
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
contourpy==1.0.7
crashtest==0.4.1
cryptography==40.0.2
cycler==0.11.0
Cython==0.29.33
diffusers==0.27.2
distlib==0.3.6
dlib==19.24.2
dulwich==0.21.3
exceptiongroup==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fb-idb==1.1.7
filelock==3.10.0
firebase-admin==6.1.0
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.39.4
fsspec==2023.6.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gradio-client==0.15.0
grpcio==1.54.0
grpcio-status==1.54.0
grpclib==0.4.5
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
hyperframe==6.0.1
idna==3.4
imgurpython==1.1.7
importlib-metadata==6.1.0
importlib-resources==5.12.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
kiwisolver==1.4.4
lockfile==0.12.2
MarkupSafe==2.1.2
matplotlib==3.7.1
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.2.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pexpect==4.8.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.1
poetry-core==1.5.2
poetry-plugin-export==1.3.0
proto-plus==1.22.2
protobuf==4.23.0
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyopenjtalk==0.3.0
pyparsing==3.0.9
pyproject-hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
PyYAML==6.0.1
rapidfuzz==2.13.7
regex==2024.4.16
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
safetensors==0.4.3
schedule==1.2.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.12
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.11.6
torch==2.2.2
torchvision==0.17.2
tqdm==4.65.0
transformers==4.40.0
treelib==1.6.4
trove-classifiers==2023.3.9
tweepy==4.14.0
typing-extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.2.3
xattr==0.10.1
zipp==3.15.0
